separate openAdvanced tests
