+++ test/open-test.js
@@ -67,44 +67,6 @@ module.exports.open = function (test, testCommon) {
   })
 }
 
-module.exports.openAdvanced = function (test, testCommon) {
-  test('test database open createIfMissing:false', function (t) {
-    var db = testCommon.factory()
-    var async = false
-
-    db.open({ createIfMissing: false }, function (err) {
-      t.ok(err, 'error')
-      t.ok(/does not exist/.test(err.message), 'error is about dir not existing')
-      t.ok(async, 'callback is asynchronous')
-      t.end()
-    })
-
-    async = true
-  })
-
-  test('test database open errorIfExists:true', function (t) {
-    var db = testCommon.factory()
-
-    db.open({}, function (err) {
-      t.error(err)
-      db.close(function (err) {
-        t.error(err)
-
-        var async = false
-
-        db.open({ createIfMissing: false, errorIfExists: true }, function (err) {
-          t.ok(err, 'error')
-          t.ok(/exists/.test(err.message), 'error is about already existing')
-          t.ok(async, 'callback is asynchronous')
-          t.end()
-        })
-
-        async = true
-      })
-    })
-  })
-}
-
 module.exports.tearDown = function (test, testCommon) {
   test('tearDown', testCommon.tearDown)
 }
@@ -114,6 +76,5 @@ module.exports.all = function (test, testCommon) {
   module.exports.setUp(test, testCommon)
   module.exports.args(test, testCommon)
   module.exports.open(test, testCommon)
-  module.exports.openAdvanced(test, testCommon)
   module.exports.tearDown(test, testCommon)
 }
